backend/models/ttt/requirements.txt
grpcio
grpcio-tools
loguru
faiss-cpu